Moving react-next messageBar to react-internal
