--- apps/perf-test/src/scenarios/MessageBarNext.tsx
@@ -1,6 +1,0 @@
-import * as React from 'react';
-import { MessageBar } from '@fluentui/react-next';
-
-const Scenario = () => <MessageBar />;
-
-export default Scenario;
