+++ packages/react-internal/src/components/MessageBar/MessageBar.types.ts
@@ -12,7 +12,7 @@ export interface IMessageBar {}
 /**
  * {@docCategory MessageBar}
  */
-export interface IMessageBarProps extends React.HTMLAttributes<HTMLElement> {
+export interface IMessageBarProps extends React.HTMLAttributes<HTMLElement>, React.RefAttributes<HTMLDivElement> {
   /**
    * Optional callback to access the IMessageBar interface. Use this instead of ref for accessing
    * the public methods and properties of the component.
