--- packages/react-next/src/components/MessageBar/index.ts
@@ -1,3 +1,0 @@
-export * from './MessageBar';
-export * from './MessageBar.base';
-export * from './MessageBar.types';
